Extract reusable "FeedStoreSpecs" methods to a shared scope so it can be used by other "FeedStore" implementation
