+++ FeedLoaderTests/Feed cache/XCTestCase+FeedStoreSpecs.swift
@@ -1,0 +1,62 @@
+//
+//  File.swift
+//  FeedLoaderTests
+//
+//  Created by AmritPandey on 20/06/22.
+//
+
+import XCTest
+@testable import FeedLoader
+
+extension FeedStoreSpecs where Self: XCTestCase {
+    @discardableResult
+    func insert(_ cache: (feed: [LocalFeedImage], timestamp: Date), to sut: FeedStore) -> Error? {
+        let exp = expectation(description: "Wait for cache insertion")
+        var insertionError: Error?
+        sut.insertCache(with: cache.feed, timestamp: cache.timestamp) { error in
+            insertionError = error
+            exp.fulfill()
+        }
+        wait(for: [exp], timeout: 1.0)
+        return insertionError
+    }
+    
+    @discardableResult
+    func delete(sut: FeedStore) -> Error? {
+        let exp = expectation(description: "Wait for cache deletion")
+        var deletionError: Error?
+        sut.deleteCache { error in
+            deletionError = error
+            exp.fulfill()
+        }
+        wait(for: [exp], timeout: 1.0)
+        return deletionError
+    }
+    
+    func expect(sut: FeedStore, toRetrieveTwice expectedResult: RetrievalCachedFeedResult, file: StaticString = #file, line: UInt = #line) {
+        expect(sut: sut, toRetrieve: expectedResult, file: file, line: line)
+        expect(sut: sut, toRetrieve: expectedResult, file: file, line: line)
+    }
+    
+    func expect(sut: FeedStore, toRetrieve expectedResult: RetrievalCachedFeedResult, file: StaticString = #file, line: UInt = #line) {
+        let exp = expectation(description: "Wait for cache retieval")
+        sut.retrieve { retrivedResult in
+            switch (retrivedResult, expectedResult) {
+            case (.empty, .empty), (.failure, .failure):
+                break
+            case let (.found(retrived), .found(expected)):
+                XCTAssertEqual(retrived.feed, expected.feed)
+                XCTAssertEqual(retrived.timestamp, expected.timestamp)
+                
+            default:
+                XCTFail("Expected \(expectedResult) instead got \(retrivedResult)")
+            }
+            exp.fulfill()
+        }
+        
+        wait(for: [exp], timeout: 1.0)
+    }
+}
+
+
+
